requirements.txt

# To ensure all is installed and only them:
# (venv)$ pip freeze | xargs pip uninstall -y
# (venv)$ pip install -r requirements.txt

# Web
Flask==2.0.3
Flask-Admin==1.6.0
# dep of Flask-Security
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
# Not maintained anymore
#Flask-Security==3.0.0
flask-security-too==4.1.2
Flask-WTF==1.0.0
WTForms==3.0.1
email_validator==1.1.3
# For json parsing in generated to_back/* code.
python-dateutil==2.8.2
# For speedup in booster.py
orjson==3.6.7
# For proper file upload to back-end. urllib3 does not stream, full file is needed.
requests==2.27.1
# Latin hyphenation of taxa names
hyphenator==0.6.0
# For PrimeVue experiment
jsonpickle==2.1.0
# Typings
types-python-dateutil
types-certifi
types-requests
types-six
# QA & compilation
mypy==0.931
mypy_extensions==0.4.3
